+++ Stickdom/level1.js
@@ -3,7 +3,6 @@ class level1 extends Phaser.Scene
     constructor()
     {
         super({key: "level1"});
-         //Global variables
     }
 
     preload ()
@@ -12,13 +11,22 @@ class level1 extends Phaser.Scene
         this.load.image('sky', 'assets/sky.png');
         this.load.image('ground', 'assets/platform.png');
         this.load.image('star', 'assets/star.png');
-        this.load.image('bomb', 'assets/bomb.png');
+        this.load.image('fireball', 'assets/fireball.png');
         this.load.spritesheet('player', 
             'assets/stickman.png',
             { frameWidth: 32, frameHeight: 50 }
         );
         this.load.image('bullet', 'assets/bullet.png');
         this.load.image('background', 'assets/basicBack.png');
+        this.load.spritesheet('healthBar',
+            'assets/healthBar.png',
+            {frameWidth: 81, frameHeight: 22}
+        );
+        this.load.image('blank','assets/blankBox.png');
+        this.load.spritesheet('fireball', 
+        'assets/fireball.png',
+            {frameWidth: 17, frameHeight: 17}   
+        );
     }
 
    
@@ -34,13 +42,21 @@ class level1 extends Phaser.Scene
         var bulletNum=0;
         this.bulletNum=1;        
         
-        //Creating keyboard input
-        var cursors = this.input.keyboard.createCursorKeys();
         //Keycodes
         this.key_Left = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.LEFT);
         this.key_Right = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.RIGHT);
         this.key_Up = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.UP);
         this.key_Space = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.SPACE);
+        this.key_Enter = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.ENTER);
+        
+        //Starting gameplay with enter
+        this.physics.pause();
+        var startText = this.add.text(200,450,'Press Enter to start the game',{ fontSize: '32px', fill: '#000' })
+        this.input.keyboard.on('keyup_ENTER', function(event)
+        {
+            startText.setText("");
+            this.physics.resume();
+        },this);
 
         //Create platforms
         var platforms;
@@ -52,24 +68,24 @@ class level1 extends Phaser.Scene
         platforms.create(600, 400, 'ground');
         platforms.create(50, 300, 'ground');
         platforms.create(750, 220, 'ground');
+        
 
-         //Camera
-         this.cameras.main.setBounds(0,0,3200,600);
+        //Camera
+        this.cameras.main.setBounds(0,0,3200,600);
         
         //Player creation
         this.player = this.physics.add.sprite(100, 450, 'player');
         this.player.setBounce(0.2);
         this.player.setCollideWorldBounds(true);
-        // this.player.setCollideWorldBounds(false);
         this.player.body.setGravityY(300) 
         //direction player facing
-        var direction = "right";
+        var direction;
         direction = "left";
         //Player collision with platform
         this.physics.add.collider(this.player, platforms);
         //Camera follows player
         this.cameras.main.startFollow(this.player); 
- 
+
          //Create animations
         this.anims.create({
             key: 'left',
@@ -109,26 +125,65 @@ class level1 extends Phaser.Scene
             frameRate: 10,
         });
 
-        //Score creation
-        var score = 0;
-        var scoreText;
-        
-        scoreText = this.add.text(16, 16, 'score: 0', { fontSize: '32px', fill: '#000' });
-        
+        //Health bar
+        this.anims.create({
+            key: 'heartsThree',
+            frames: [{key: 'healthBar', frame: 0}],
+            frameRate: 10,
+        });
+
+        this.anims.create({
+            key: 'heartsTwo',
+            frames: [{key: 'healthBar', frame: 1}],
+            frameRate: 10,
+        });
+
+        this.anims.create({
+            key: 'heartsOne',
+            frames: [{key: 'healthBar', frame: 2}],
+            frameRate: 10,
+        });
+
+        this.anims.create({
+            key: 'heartsZero',
+            frames: [{key: 'healthBar', frame: 3}],
+            frameRate: 10,
+        });
+
+        //Fireball animation
+        this.anims.create({
+            key: 'fireballMovement',
+            frames: this.anims.generateFrameNumbers('fireball', { start: 0, end: 1 }),
+            frameRate: 10,
+            repeat: -1
+        });
+
+        //Healthbar
+        this.health = this.physics.add.group();
+        this.healthBar = this.health.create(40,16,'blank');
+        this.healthBar.anims.play('heartsThree');
+        this.healthBar.setCollideWorldBounds(true);
+        //Collider so healthbar doesn't go off screen
+        this.physics.add.collider(this.healthBar, platforms);
+        platforms.create(16,16,'blank');
+        //Health count
+        var healthCount = 3;
+        this.healthCount = 3;
+
         //adding enemies
-        var bombs = this.physics.add.group({
+        var fireballs = this.physics.add.group({
             gravityY: 300,
         });
-        this.physics.add.collider(bombs, platforms);
-        this.physics.add.collider(this.player, bombs, hitBomb, null, this);
+        this.physics.add.collider(fireballs, platforms);
+        this.physics.add.collider(this.player, fireballs, hitfireball, null, this);
 
-        //hit bomb FUNCTION!!!!!!
-        function hitBomb (player, bomb)
+        //Player hits fireball
+        function hitfireball (player, fireball)
         {
-            this.physics.pause();
-            player.setTint(0xff0000);
+            // this.physics.pause();
             player.anims.play('turn');
-            this.gameOver = true;
+            this.healthCount--;
+            fireball.disableBody(true, true);
         }
 
         //Creating group of stars
@@ -142,7 +197,6 @@ class level1 extends Phaser.Scene
         stars.children.iterate(function (child) {
 
             child.setBounceY(Phaser.Math.FloatBetween(0.4, 0.8));
-
         });
         
         //Check star collision with ground
@@ -155,16 +209,14 @@ class level1 extends Phaser.Scene
         function collectStar (player, star)
         {
             star.disableBody(true, true);
-
-            score += 10;
-            scoreText.setText('Score: ' + score);
-
             var x = (player.x < 400) ? Phaser.Math.Between(400, 800) : Phaser.Math.Between(0, 400);
-            var bomb = bombs.create(x, 16, 'bomb');
-            bomb.setBounce(1);
-            bomb.setCollideWorldBounds(true);
-            bomb.setVelocity(Phaser.Math.Between(-200, 200), 20);
-            bomb.allowGravity = false;
+            var fireball = fireballs.create(x, 16, 'fireball');
+            // console.log(fireball);
+            fireball.anims.play('fireballMovement', true);
+            fireball.setBounce(1);
+            fireball.setCollideWorldBounds(true);
+            fireball.setVelocity(Phaser.Math.Between(-200, 200), 20);
+            fireball.allowGravity = false;
                 
             if (stars.countActive(true) === 0)
             {
@@ -176,28 +228,25 @@ class level1 extends Phaser.Scene
 
             }
         }
+        
+        //add bullets group
+        this.bullets = this.physics.add.group();
+        this.physics.add.collider(this.bullets, fireballs, bulletHit, null, this);
+        this.physics.add.collider(this.bullets, platforms, bulletBounds, null, this);
+        this.physics.add.collider(this.bullets,this.bullets, bulletTouchingBullet, null, this);
 
         //If bullet hit disable
-        function bulletHit (bullet, bomb)
+        function bulletHit (bullet, fireball)
         {
             bullet.disableBody(true,true);
-            bomb.disableBody(true,true);
+            fireball.disableBody(true,true);
         }
 
         //Delete a bullet when hitbox overlap
         function bulletTouchingBullet (bulletOne,bulletTwo)
         {
             this.bulletTwo.disableBody(true,true);
-
         }
-        
-        //add bullets group
-        this.bullets = this.physics.add.group();
-        this.physics.add.collider(this.bullets, bombs, bulletHit, null, this);
-        this.physics.add.collider(this.bullets, platforms, bulletBounds, null, this);
-        this.physics.add.collider(this.bullets,this.bullets, bulletTouchingBullet, null, this);
-        
-        
         function bulletBounds (bullet, platforms)
         {
             bullet.disableBody(true,true);
@@ -211,6 +260,7 @@ class level1 extends Phaser.Scene
         if(this.key_Space.isDown)
         {
             this.player.setVelocityX(0);
+            this.healthBar.setVelocityX(0);
             if(this.direction=== "left")
             {
                 this.player.anims.play('shootLeft');
@@ -218,15 +268,13 @@ class level1 extends Phaser.Scene
             else if(this.direction === "right")
             {
                 this.player.anims.play('shootRight');
+            }
 
-
-            } 
         }
         //set shooting function
         else if(this.key_Space._justUp)
         //Just up is used if the button is pressed, hence just popped up
         {
-            // isShooting === true;
             var velX = 200;
             if(this.direction=== "left")
             {
@@ -255,6 +303,7 @@ class level1 extends Phaser.Scene
                 this.bulletNum = 1;
             }
             this.key_Space._justUp = false;
+            // console.log(this.healthCount);
         }
         //Moving left
         else if (this.key_Left.isDown)
@@ -262,6 +311,10 @@ class level1 extends Phaser.Scene
             this.player.setVelocityX(-160);
             this.player.anims.play('left', true);
             this.direction = "left";
+            if(this.player.x>=395)
+            {
+                this.healthBar.setVelocityX(-160);
+            }
         }
         //Moving right
         else if (this.key_Right.isDown)
@@ -270,11 +323,16 @@ class level1 extends Phaser.Scene
 
             this.player.anims.play('right', true);
             this.direction = "right";
+            if(this.player.x>=395 ) 
+            {
+                this.healthBar.setVelocityX(160);
+            }
         }
         //Not moving
         else
         {
             this.player.setVelocityX(0);
+            this.healthBar.setVelocityX(0);
             if(this.direction === "left")
             {
                 this.player.anims.play('faceLeft');
@@ -282,9 +340,8 @@ class level1 extends Phaser.Scene
             else if(this.direction === "right")
             {
                 this.player.anims.play('faceRight');
-            }            
+            } 
         }
-        
 
         //Jump 
         if (this.key_Up.isDown && this.player.body.touching.down)
@@ -292,14 +349,34 @@ class level1 extends Phaser.Scene
             this.player.setVelocityY(-325);
         }
 
+        //Different heart animations for different health
+        if(this.healthCount === 2)
+        {
+            this.healthBar.anims.play('heartsTwo');
+        }
+        else if (this.healthCount === 1)
+        {
+            this.healthBar.anims.play('heartsOne');
+        }
+        else if (this.healthCount === 0)
+        {
+            this.healthBar.anims.play('heartsZero');
+            this.gameOver = true;
+            this.physics.pause();
+        }
+
         //RESET DEBUGGING USE ONLY
         this.key_R = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.R);
-        if (this.key_R.isDown)
+        if (this.key_R._justUp)
         {
             this.physics.resume();
             // this.player.setTint();
             this.player.anims.play('turn');
             this.gameOver = false;
+            this.healthCount = 3;
+            this.healthBar.anims.play('heartsThree');
+            this.key_R._justUp = false;
         }
     }
+    
 }