+++ CHANGELOG.md
@@ -16,3 +16,6 @@
 	- Basic health bar on top
 	- Created a blank box to make sure health doesn't go off platform
 	- Made health bar velocity based on player movement
+	- Arrows have gravity effect
+- 4/20/2018: Game starts on Enter
+